@@ -3,20 +3,16 @@ using System.Net;
 using System.Text;
 using System.Text.Json;
 
-namespace ABCRetailersFunctions.Helpers
-{
+namespace ABCRetailersFunctions.Helpers;
+
     public static class HttpJson
     {
-        public static readonly JsonSerializerOptions _options = new()
-        { 
-            PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            WriteIndented = true
-        };
+        public static readonly JsonSerializerOptions _json = new(JsonSerializerDefaults.Web);
 
         public static async Task<T> ReadAsync<T>(HttpRequestData req)
         {
             using var s = req.Body;
-            return await JsonSerializer.DeserializeAsync<T>(s, _options);
+            return await JsonSerializer.DeserializeAsync<T>(s, _json);
         }
         public static HttpResponseData OK<T>(HttpRequestData req, T body)
             => Write(req, HttpStatusCode.OK, body);
@@ -34,13 +30,7 @@ namespace ABCRetailersFunctions.Helpers
         }
 
         public static HttpResponseData NotFound(HttpRequestData req, string message = "Not Found")
-        {
-            var r = req.CreateResponse(HttpStatusCode.NotFound);
-            r.Headers.Add("Content-Type", "application/json");
-            string json = JsonSerializer.Serialize(new { error = message }, _options);
-            r.WriteString(json);
-            return r;
-        }
+            => Text(req, HttpStatusCode.NotFound, message);
 
         public static HttpResponseData Text(HttpRequestData req, HttpStatusCode code, string message)
         {
@@ -54,9 +44,8 @@ namespace ABCRetailersFunctions.Helpers
         {
             var r = req.CreateResponse(code);
             r.Headers.Add("Content-Type", "application/json; charset=utf-8");
-            var json = JsonSerializer.Serialize(body, _options);
+            var json = JsonSerializer.Serialize(body, _json);
             r.WriteString(json, Encoding.UTF8);
             return r;
         }
     }
-}
