@@ -3,14 +3,10 @@ using ABCRetailersFunctions.Helpers;
 using ABCRetailersFunctions.Models;
 using Azure.Data.Tables;
 using Azure.Storage.Blobs;
-using Azure.Storage.Queues;
-using Azure.Storage.Queues.Models;
-using Microsoft.AspNetCore.Authentication.OAuth.Claims;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Configuration;
-
+using static Grpc.Core.Metadata;
 
 namespace ABCRetailersFunctions.Functions;
 
@@ -22,7 +18,7 @@ public class ProductsFunctions
 
     public ProductsFunctions(IConfiguration con)
     {
-        _conn = con["STORAGE_CONNECTION"] ?? throw new InvalidOperationException("Missing STORAGE_CONNECTION");
+        _conn = con["STORAGE_CONNECTION"] ?? throw new InvalidOperationException("STORAGE_CONNECTION missing");
         _table = con["TABLE_PRODUCT"] ?? "Product";
         _images = con["BLOB_PRODUCT_IMAGES"] ?? "product-images";
     }
@@ -40,9 +36,7 @@ public class ProductsFunctions
             items.Add(Map.ToDto(e));
         }
 
-        var response = req.CreateResponse(System.Net.HttpStatusCode.OK);
-        await response.WriteAsJsonAsync(items);
-        return response;
+        return HttpJson.OK(req, items);
     }
 
     [Function("Products_Get")]
@@ -57,26 +51,22 @@ public class ProductsFunctions
         try
         {
             var entity = await table.GetEntityAsync<ProductEntity>("Product", id);
-            var response = req.CreateResponse(System.Net.HttpStatusCode.OK);
-            await response.WriteAsJsonAsync(Map.ToDto(entity.Value));
-            return response;
+            return HttpJson.OK(req, Map.ToDto(entity.Value));
         }
-        catch (Azure.RequestFailedException ex) when (ex.Status == 404)
+        catch
         {
-            var response = req.CreateResponse(System.Net.HttpStatusCode.NotFound);
-            await response.WriteStringAsync("Product not found");
-            return response;
+            return HttpJson.NotFound(req, "Product not Found");
         }
     }
 
     [Function("Product_Create")]
 
     public async Task<HttpResponseData> Create(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "products")] HttpRequestData req
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "products")] HttpRequestData req, FunctionContext ctx
         )
     {
 
-        var contentType = req.Headers.TryGetValues("Content-Type", out var ct) ? ct.ToString() : null;
+        var contentType = req.Headers.TryGetValues("Content-Type", out var ct) ? ct.First() : "";
         var table = new TableClient(_conn, _table);
         await table.CreateIfNotExistsAsync();
 
@@ -84,13 +74,13 @@ public class ProductsFunctions
         double price = 0;
         int stock = 0;
 
-        if (!string.IsNullOrWhiteSpace(contentType) && contentType.StartsWith("multipart/form-data", StringComparison.OrdinalIgnoreCase))
+        if (contentType.StartsWith("multipart/form-data", StringComparison.OrdinalIgnoreCase))
         {
             var form = await MultipartHelper.ParseAsync(req.Body, contentType);
             name = form.Text.GetValueOrDefault("ProductName") ?? "";
             desc = form.Text.GetValueOrDefault("Description") ?? "";
-            double.TryParse(form.Text.GetValueOrDefault("Price"), out price);
-            int.TryParse(form.Text.GetValueOrDefault("StockAvailable"), out stock);
+            double.TryParse(form.Text.GetValueOrDefault("Price") ?? "0", out price);
+            int.TryParse(form.Text.GetValueOrDefault("StockAvailable") ?? "0", out stock);
 
             var file = form.Files.FirstOrDefault(f => f.FileName == "ImageFile");
             if (file != null && file.Data.Length > 0)
@@ -112,22 +102,16 @@ public class ProductsFunctions
             var body = await HttpJson.ReadAsync<Dictionary<string, object>>(req) ?? new();
             name = body.TryGetValue("ProductName", out var pn) ? pn?.ToString() ?? "" : "";
             desc = body.TryGetValue("Description", out var d) ? d?.ToString() ?? "" : "";
-            price = body.TryGetValue("Price", out var pr) && double.TryParse(pr?.ToString(), out var p) ? p : 0;
-            stock = body.TryGetValue("StockAvailable", out var st) && int.TryParse(st?.ToString(), out var s) ? s : 0;
+            price = body.TryGetValue("Price", out var pr) ? Convert.ToDouble(pr) : 0;
+            stock = body.TryGetValue("StockAvailable", out var st) ? Convert.ToInt32(st) : 0;
             imageUrl = body.TryGetValue("ImageUrl", out var iu) ? iu?.ToString() ?? "" : "";
         }
 
         if (string.IsNullOrWhiteSpace(name))
-        {
-            var badResponse = req.CreateResponse(System.Net.HttpStatusCode.BadRequest);
-            await badResponse.WriteStringAsync("ProductName is required");
-            return badResponse;
-        }
+            return HttpJson.Bad(req, "ProductName is required");
 
         var entity = new ProductEntity
         {
-            PartitionKey = "Product",
-            RowKey = Guid.NewGuid().ToString(),
             ProductName = name,
             Description = desc,
             Price = price,
@@ -137,9 +121,7 @@ public class ProductsFunctions
 
         await table.AddEntityAsync(entity);
 
-        var createdResponse = req.CreateResponse(System.Net.HttpStatusCode.Created);
-        await createdResponse.WriteAsJsonAsync(Map.ToDto(entity));
-        return createdResponse;
+        return HttpJson.Created(req, Map.ToDto(entity));
 
     }
 
@@ -148,7 +130,7 @@ public class ProductsFunctions
     [HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "products/{id}")] HttpRequestData req,
     string id)
     {
-        var contentType = req.Headers.TryGetValues("Content-Type", out var ctValues) ? ctValues.FirstOrDefault() : null;
+        var contentType = req.Headers.TryGetValues("Content-Type", out var ct) ? ct.First() : "";
         var table = new TableClient(_conn, _table);
 
         try
@@ -156,7 +138,7 @@ public class ProductsFunctions
             var response = await table.GetEntityAsync<ProductEntity>("Product", id);
             var entity = response.Value;
 
-            if (!string.IsNullOrWhiteSpace(contentType) && contentType.StartsWith("multipart/form-data", StringComparison.OrdinalIgnoreCase))
+            if (contentType.StartsWith("multipart/form-data", StringComparison.OrdinalIgnoreCase))
             {
                 var form = await MultipartHelper.ParseAsync(req.Body, contentType);
 
@@ -165,31 +147,46 @@ public class ProductsFunctions
                 if (form.Text.TryGetValue("Price", out var pr) && double.TryParse(pr, out var price)) entity.Price = price;
                 if (form.Text.TryGetValue("StockAvailable", out var st) && int.TryParse(st, out var stock)) entity.StockAvailable = stock;
                 if (form.Text.TryGetValue("ImageUrl", out var imageUrl)) entity.ImageUrl = imageUrl;
+
+                var file = form.Files.FirstOrDefault(f => f.FieldName == "ImageFile");
+                if (file != null && file.Data.Length > 0)
+                {
+
+                    var container = new BlobContainerClient(_conn, _images);
+                    await container.CreateIfNotExistsAsync();
+                    var blob = container.GetBlobClient($"{Guid.NewGuid():N}-{file.FileName}");
+                    await using var s = file.Data;
+                    await blob.UploadAsync(s, overwrite: false);
+                    entity.ImageUrl = blob.Uri.ToString();
+                }
             }
             else
             {
-                var body = await HttpJson.ReadAsync<Dictionary<string, object>>(req);
-                if (body is not null)
-                {
-                    if (body.TryGetValue("ProductName", out var pn)) entity.ProductName = pn?.ToString() ?? entity.ProductName;
-                    if (body.TryGetValue("Description", out var d)) entity.Description = d?.ToString() ?? entity.Description;
-                    if (body.TryGetValue("Price", out var pr) && double.TryParse(pr?.ToString(), out var price)) entity.Price = price;
-                    if (body.TryGetValue("StockAvailable", out var st) && int.TryParse(st?.ToString(), out var stock)) entity.StockAvailable = stock;
-                    if (body.TryGetValue("ImageUrl", out var iu)) entity.ImageUrl = iu?.ToString() ?? entity.ImageUrl;
-                }
+                var body = await HttpJson.ReadAsync<Dictionary<string, object>>(req) ?? new();
+                if (body.TryGetValue("ProductName", out var pn)) entity.ProductName = pn?.ToString() ?? entity.ProductName;
+                if (body.TryGetValue("Description", out var d)) entity.Description = d?.ToString() ?? entity.Description;
+                if (body.TryGetValue("Price", out var pr) && double.TryParse(pr?.ToString(), out var price)) entity.Price = price;
+                if (body.TryGetValue("StockAvailable", out var st) && int.TryParse(st?.ToString(), out var stock)) entity.StockAvailable = stock;
+                if (body.TryGetValue("ImageUrl", out var iu)) entity.ImageUrl = iu?.ToString() ?? entity.ImageUrl;
+
             }
 
             await table.UpdateEntityAsync(entity, entity.ETag, TableUpdateMode.Replace);
 
-            var okResponse = req.CreateResponse(System.Net.HttpStatusCode.OK);
-            await okResponse.WriteAsJsonAsync(Map.ToDto(entity));
-            return okResponse;
+            return HttpJson.OK(req, Map.ToDto(entity));
         }
-        catch (Azure.RequestFailedException ex) when (ex.Status == 404)
+        catch
         {
-            var notFoundResponse = req.CreateResponse(System.Net.HttpStatusCode.NotFound);
-            await notFoundResponse.WriteStringAsync("Product not found");
-            return notFoundResponse;
+            return HttpJson.NotFound(req, "Product not found");
         }
+    }
+
+    [Function("Products_Delete")]
+    public async Task<HttpResponseData> Delete(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "products/{id}")] HttpRequestData req, string id)
+    {
+        var table = new TableClient(_conn, _table);
+        await table.DeleteEntityAsync("Product", id);
+        return HttpJson.NoContent(req);
     }
 }
