@@ -1,10 +1,9 @@
-using ABCRetails.Functions.Helpers;
+using ABCRetailersFunctions.Helpers;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Configuration;
 using Azure.Storage.Blobs;
 using Azure.Storage.Files.Shares;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace ABCRetails.Functions.Functions;
 
@@ -18,7 +17,7 @@ public class UploadsFunctions
 
     public UploadsFunctions(IConfiguration con)
     {
-        _conn = con["STORAGE_CONNECTION"] ?? throw new InvalidOperationException();
+        _conn = con["STORAGE_CONNECTION"] ?? throw new InvalidOperationException("STORAGE_CONNECTION missing");
         _proofs = con["BLOB_PAYMENT_PROOFS"] ?? "payment-proofs";
         _share = con["FILESHARE_CONTRACTS"] ?? "contracts";
         _shareDir = con["FILESHARE_DIR_PAYMENTS"] ?? "payments";
@@ -27,15 +26,16 @@ public class UploadsFunctions
 
     [Function("Uploads_ProofOfPayment")]
     public async Task<HttpResponseData> Proof(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "uploads/proof-")] HttpRequest req)
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "uploads/proof-of-payment")] HttpRequestData req)
     {
-        var contentType = req.Headers.TryGetValues("Content-Type", out var ct) ? ct : "application/json";
+        var contentType = req.Headers.TryGetValues("Content-Type", out var values) ? values.First() : "";
+
         if (!contentType.StartsWith("multipart/form-data", StringComparison.OrdinalIgnoreCase))
-            return HttpJsonServiceExtensions.Bad(req, "Expected multipart/form-data");
+            return HttpJson.Bad(req, "Expected multipart/form-data");
 
         var form = await MultipartHelper.ParseAsync(req.Body, contentType);
-        var file = form.Files.FirstOrDefault(form => form.FieldName == "ProofOfPayment");
-        if (file is null || file.Data.Length == 0) return HttpJsonServiceExtensions.Bad(req, "Proof");
+        var file = form.Files.FirstOrDefault(f => f.FieldName == "ProofOfPayment");
+        if (file is null || file.Data.Length == 0) return HttpJson.Bad(req, "ProofOfPayment file is required");
 
         var orderId = form.Text.GetValueOrDefault("OrderId");
         var customerName = form.Text.GetValueOrDefault("CustomerName");
@@ -55,12 +55,12 @@ public class UploadsFunctions
         await dir.CreateIfNotExistsAsync();
 
         var fileClient = dir.GetFileClient(blobName + ".txt");
-        var meta = $"UploadedAtUtc: {DateTimeOffset.UtcNow:O}\nOrderId: {orderId}\n";
+        var meta = $"UploadedAtUtc: {DateTimeOffset.UtcNow:O}\nOrderId: {orderId}\nCustomerName: {customerName}\nBlobUrl: {blob.Uri}";
         var bytes = System.Text.Encoding.UTF8.GetBytes(meta);
         using var ms = new MemoryStream(bytes);
         await fileClient.CreateAsync(ms.Length);
         await fileClient.UploadAsync(ms);
 
-        return HttpJson.Ok(req, new { fileName = blobName, blobUrl = blob.Uri.ToString()});
+        return HttpJson.OK(req, new { fileName = blobName, blobUrl = blob.Uri.ToString()});
     }
 }
