@@ -1,43 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
+﻿
+namespace ABCRetailersFunctions.Models;
 
-namespace ABCRetailersFunctions.Models
-{
-    public class CustomerDto
-    {
-        public string CustomerId { get; set; } = string.Empty;
-        public string Name { get; set; } = string.Empty;
-        public string Surname { get; set; } = string.Empty;
-        public string Username { get; set; } = string.Empty;
-        public string Email { get; set; } = string.Empty;
-        public string ShippingAddress { get; set; } = string.Empty;
-    }
-    public class OrderDto
-    {
-        public string OrderId { get; set; } = string.Empty;
-        public string CustomerId { get; set; } = string.Empty;
-        public string Username { get; set; } = string.Empty;
-        public string ProductId { get; set; } = string.Empty;
-        public string ProductName { get; set; } = string.Empty;
-        public DateTimeOffset OrderDate { get; set; }
-        public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
-        public string Status { get; set; } = "Submitted";
-    }
+    public record CustomerDto
+        (string Id, string Name, string Surname, string Username, string Email, string ShippingAddress);
+    public record OrderDto
+        (string Id, string CustomerId, string ProductId, string ProductName, int Quantity, decimal UnitPrice, decimal TotalAmount, DateTimeOffset OrderDateUtc, string Status);
 
-    public class ProductDto
-    {
-        public string ProductId { get; set; } = string.Empty;
-        public string ProductName { get; set; } = string.Empty;
-        public string Description { get; set; } = string.Empty;
-        public double Price { get; set; }
-        public int StockAvailable { get; set; }
-        public string ImageUrl { get; set; } = string.Empty;
-    }
-}
+    public record ProductDto
+        (string Id, string ProductName, string Description, decimal Price, int StockAvailable, string ImageUrl);
