@@ -6,17 +6,13 @@ namespace ABCRetails.Controllers
 {
     public class ProductController : Controller
     {
-        private readonly IAzureStorageService _storageService;
-        private readonly ILogger<ProductController> _logger;
+        private readonly IFunctionApi _api;
 
-        public ProductController(IAzureStorageService storageService, ILogger<ProductController> logger)
-        {
-            _storageService = storageService;
-            _logger = logger;
-        }
+        public ProductController(IFunctionApi api) => _api = api;
+
         public async Task<IActionResult> Index()
         {
-            var products = await _storageService.GetAllEntitiesAsync<Product>();
+            var products = await _api.GetProductsAsync();
             return View(products);
         }
 
@@ -25,136 +21,57 @@ namespace ABCRetails.Controllers
             return View();
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
+        [HttpPost, ValidateAntiForgeryToken]
 
         public async Task<IActionResult> Create(Product product, IFormFile? imageFile)
         {
-            if (Request.Form.TryGetValue("Price", out var priceFormValue)) //price parsing
+
+            if (!ModelState.IsValid) return View(product);
+                        
+            try
             {
-                _logger.LogInformation("Raw price from form: '{PriceFormValue}'", priceFormValue.ToString());
-
-                if (double.TryParse(priceFormValue, out var parsedPrice))
-                {
-
-                    product.Price = parsedPrice;
-                    product.PriceString = parsedPrice.ToString("F2");
-                    _logger.LogInformation("Successfully parsed price: {Price}", parsedPrice);
-
-                }
-                else
-                {
-
-                    _logger.LogWarning("Failed to parse price: {PriceFormValue}", priceFormValue.ToString());
-                }
+                var saved = await _api.CreateProductAsync(product, imageFile);
+                
+                TempData["Success"] = $"Product '{saved.ProductName}' created successfully with price {saved.Price:C}!";
+                
+                return RedirectToAction(nameof(Index));
             }
-
-            _logger.LogInformation("Final product price: {Price}", product.Price);
-
-            if(ModelState.IsValid)
+            
+            catch (Exception ex)
             {
-                try
-                {
-                    if (product.Price <= 0) 
-                    {
-                        ModelState.AddModelError("Price", "Price must be greater than $0.00");
-                        return View(product);
-                    }
-
-                    //Upload image
-                    if (imageFile != null && imageFile.Length > 0) 
-                    {
-                        var imageUrl = await _storageService.UploadImageAsync(imageFile, "product-images");
-                        product.ImageUrl = imageUrl;
-                    }
-
-                    await _storageService.AddEntityAsync(product);
-                    TempData["Success"] = $"Product '{product.ProductName}' created successfully with price {product.Price:C}!";
-                    return RedirectToAction(nameof(Index));
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogInformation(ex, "Error creating product");
-                    ModelState.AddModelError("", $"Error creating product: {ex.Message}");
-
-                }
+                ModelState.AddModelError("", $"Error creating product: {ex.Message}");
+                
+                return View(product);
             }
-            return View(product);
         }
 
         public async Task<IActionResult> Edit(string id) //Edit action
         {
-            if (string.IsNullOrEmpty(id))
-            {
-                return NotFound();
-            }
+            if (string.IsNullOrWhiteSpace(id)) return NotFound();
 
-            var product = await _storageService.GetEntityAsync<Product>("Product", id);
-            if (product == null)
-            {
-                return NotFound();
-            }
-
-            return View(product);
+            var product = await _api.GetProductAsync(id);
+ 
+            return product == null ? NotFound() : View(product);
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
+        [HttpPost, ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(Product product, IFormFile? imageFile)
         {
 
-            if (Request.Form.TryGetValue("Price", out var priceFormValue)) //price parsing
+            if (!ModelState.IsValid) return View(product);
+            
+            try
             {
 
-                if (double.TryParse(priceFormValue, out var parsedPrice))
-                {
-
-                    product.Price = parsedPrice;
-                    _logger.LogInformation("Edit: Successfully parsed price: {Price}", parsedPrice);
-
-                }
-                
+                var updated = await _api.UpdateProductAsync(product.ProductId, product, imageFile);
+                TempData["Success"] = $"Product '{updated.ProductName}' updated successfully!";
+                return RedirectToAction(nameof(Index));
             }
-
-            if (ModelState.IsValid)
+            catch (Exception ex)
             {
-                try
-                {
-                    //get original product for ETag
-                    var originalProduct = await _storageService.GetEntityAsync<Product>("Product", product.RowKey);
-                    
-                    if (originalProduct == null)
-                    {
-                       
-                        return NotFound();
-                    }
-
-                    //update properties but keep the first ETag
-                    originalProduct.ProductName = product.ProductName;
-                    originalProduct.Description = product.Description;
-                    originalProduct.Price = product.Price;
-                    originalProduct.PriceString = product.Price.ToString("F2");
-                    originalProduct.StockAvailable = product.StockAvailable;
-
-                    //upload new image
-                    if (imageFile != null && imageFile.Length > 0)
-                    {
-                        var imageUrl = await _storageService.UploadImageAsync(imageFile, "product-images");
-                        originalProduct.ImageUrl = imageUrl;
-                    }
-
-                    await _storageService.UpdateEntityAsync(originalProduct);
-                    TempData["Success"] = "Product updated successfully!";
-                    return RedirectToAction(nameof(Index));
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error updated product: {Message}", ex.Message);
-                    ModelState.AddModelError("", $"Error updating product: {ex.Message}");
-
-                }
+                ModelState.AddModelError("", $"Error updating product: {ex.Message}");
+                return View(product);
             }
-            return View(product);
         }
 
         [HttpPost]
@@ -164,9 +81,8 @@ namespace ABCRetails.Controllers
 
             try
             {
-                await _storageService.DeleteEntityAsync<Product>("Product", id);
+                await _api.DeleteProductAsync(id);
                 TempData["Success"] = "Product deleted successfully!";
-                return RedirectToAction(nameof(Index));
             }
             catch (Exception ex)
             {
