@@ -6,51 +6,45 @@ namespace ABCRetails.Controllers
 {
     public class UploadController : Controller
     {
-        private readonly IAzureStorageService _storageService;
+        private readonly IFunctionApi _api;
 
-        public UploadController(IAzureStorageService storageService)
-        {
-            _storageService = storageService;
-        }
+        public UploadController(IFunctionApi api) => _api = api;
+
         public IActionResult Index()
         {
             return View(new FileUploadModel());
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
+        [HttpPost, ValidateAntiForgeryToken]
 
         public async Task<IActionResult> Index(FileUploadModel model)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid) return View(model);
+            
+            try
             {
-                try
+                if (model.ProofOfPayment == null || model.ProofOfPayment.Length > 0)
                 {
-                    if (model.ProofOfPayment != null && model.ProofOfPayment.Length > 0)
-                    {
-                        //upload to blob storage
-                        var fileName = await _storageService.UploadFileAsync(model.ProofOfPayment, "payment-proofs");
-
-                        //Upload to file share for contracts
-                        await _storageService.UploadToFileShareAsync(model.ProofOfPayment, "contracts", "payments");
-
-                        TempData["Success"] = $"File upload successfully! File name: {fileName}";
-
-                        return View(new FileUploadModel());
-                    }
-                    else 
-                    {
-                        ModelState.AddModelError("ProofOfPayment", "Please select a file to upload");
-                    }
-
-                    
+                    ModelState.AddModelError("ProofOfPayment", "Please select a file to upload");
+                    return View(model);
                 }
-                catch (Exception ex)
-                {
-                    ModelState.AddModelError("", $"Error creating customer: {ex.Message}");
-                }
+
+                var fileName = await _api.UploadProofOfPaymentAsync(
+                    model.ProofOfPayment,
+                    model.OrderId,
+                    model.CustomerName
+                );
+
+                TempData["Success"] = $"File upload successfully! File name: {fileName}";
+
+                return View(new FileUploadModel());
+
             }
-            return View(model);
+            catch (Exception ex)
+            {
+                ModelState.AddModelError("", $"Error creating customer: {ex.Message}");
+                return View(model);
+            }
         }
     }
 }
